% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_rprofile.R
\name{clear_rprofile}
\alias{clear_rprofile}
\title{A function to update my .Rprofile from a version on Github}
\usage{
clear_rprofile(local.rprofile.path = "~/.Rprofile")
}
\arguments{
\item{local.rprofile.path}{character value.  Path to local .Rprofile.  Defaults to
common "~/.Rprofile".}
}
\description{
A function to update my .Rprofile from a version on Github
}
\details{
This function takes an existing Rprofile file and deletes any cusomization.
    Note that it checks for the file at the common "~/.Rprofile" location, but can be
    manually adjusted for projects, etc.
}
\examples{
library(learylib)

clear_rprofile()  # using default settings

}
