% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_clipboard.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{A function to update my .Rprofile from a version on Github}
\usage{
copy_to_clipboard(data.input, row.names.input = FALSE)
}
\arguments{
\item{data.input}{data.frame. Data frame desired to by moved outside of R}

\item{row.names.input}{logical. Should row names be included in in the output
data frame.  Defaults to FALSE.}
}
\description{
A function to update my .Rprofile from a version on Github
}
\details{
This function takes a data frame and copies it to the computer clipboard,
    allowing user to paste the data into excel, word, email, etc.
}
\examples{
library(learylib)

copy_to_clipboard(mtcars)

}
