% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leary_html.R
\name{leary_html}
\alias{leary_html}
\title{Personalized rmarkdow::html_document template}
\usage{
leary_html(toc = TRUE, ...)
}
\arguments{
\item{toc}{if TRUE, displays the table of contents}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
Personalized rmarkdow::html_document template
}
