% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{Determine if value(s) are NOT in another value or set.}
\usage{
not_in(x, y)
}
\arguments{
\item{x}{single value or vector}

\item{y}{single value or vector}
}
\description{
Determine if value(s) are NOT in another value or set.
}
\details{
This is a helper function to see if values are NOT IN other values.
    Mainly this helps me quickly check without having to negate the %in% function.
    This is a test to see if this function is helpful, and might be deleted.
    later on.
}
\examples{
library(learylib)
library(dplyr)

"test" \%>\% not_in(c("quiz", "homework"))

"test" \%>\% not_in(c("quiz", "test"))

}
