% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_na.R
\name{not_na}
\alias{not_na}
\title{Determine if value(s) are NOT NA}
\usage{
not_na(x)
}
\arguments{
\item{x}{single value or vector}
}
\description{
Determine if value(s) are NOT NA
}
\details{
This is a helper function to check for NOT NA.  This is mainly
    to hel me quickly check without having to negate 'is.na' with '!is.na(x)'.
    This is a test to see if this function is helpful, and might be deleted
    later on.
}
\examples{
library(learylib)

x <- c(1,2,3,NA)

not_na(x)

}
