% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile_clear.R
\name{rprofile_clear}
\alias{rprofile_clear}
\title{A function to update my .Rprofile from a version on Github}
\usage{
rprofile_clear(local.rprofile.path = "~/.Rprofile")
}
\arguments{
\item{local.rprofile.path}{character value.  Path to local .Rprofile.  Defaults to
common "~/.Rprofile".}
}
\description{
A function to update my .Rprofile from a version on Github
}
\details{
This function takes an existing Rprofile file and deletes the text.
    Note that it checks for the file at the common "~/.Rprofile" location, but can be
    manually adjusted for projects, etc.
}
\examples{
library(learylib)

rprofile_clear()  # using default settings

}
