% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_last_error.R
\name{search_last_error}
\alias{search_last_error}
\title{Google search R error messages}
\usage{
search_last_error(additonal.term = NULL)
}
\arguments{
\item{additonal.term}{character. Additional text to add to Google search after "R".}
}
\description{
Google search R error messages
}
\details{
This function takes the last error in my R session and creates a Google search
    for an answer. 93 percent of the time this will be to remember something I
    have done in `{ggplot2}` 1,573 times before but forgot.  For example,
    adding "ggplot" will search 'R, ggplot, {error message here}'. Heavily
    borrowed from & influenced from David Robinson's github.
}
\examples{
library(learylib)

search_last_error("ggplot")

}
