% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_lib_paths.R
\name{see_lib_paths}
\alias{see_lib_paths}
\title{Print the library path(s) from .libPaths()}
\usage{
see_lib_paths()
}
\description{
Print the library path(s) from .libPaths()
}
\details{
This is a helper to "pretty print" the library paths for my existing
    session.
}
\examples{
library(learylib)

see_lib_paths() #prints all library path(s)

}
