% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leary_utils.R
\name{see_pkg_funcs}
\alias{see_pkg_funcs}
\title{Print functions within a package}
\usage{
see_pkg_funcs(package, all.names = FALSE, pattern)
}
\arguments{
\item{package}{R package to inspect}

\item{all.names}{logical value. If TRUE, all object names are returned.
If FALSE, names which begin with a . are omitted.}

\item{pattern}{an optional regular expression. Only names matching pattern
are returned. glob2rx can be used to convert wildcard patterns to regular
expressions.}
}
\description{
Print functions within a package
}
\details{
This is a helper to let me see the different functions offered by a
   package. NOTE: The package most be installed & loaded (i.e. library(janitor))
}
\examples{
library(learylib)

see_pkg_funcs(janitor)


}
