% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_from_github}
\alias{text_from_github}
\title{A function to read text from github (or really any site)}
\usage{
text_from_github(url)
}
\arguments{
\item{url}{URL to RAW github text file}
}
\description{
A function to read text from github (or really any site)
}
\details{
This is a helper function for getting text files from my Github repo.
    It is currently used in the update_rprofile function and could be used in
    future functions such as update_gitignore.
}
