% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_leary.R
\name{theme_leary}
\alias{theme_leary}
\title{Minimal ggplot2 theme}
\usage{
theme_leary(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{Other arguments passed to \code{theme_minimal}}
}
\description{
Minimal ggplot2 theme
}
\details{
This is a draft veresion and will be updated as I improve my package
    development skills.  It does not currently match the Roboto font used in
    my rmarkdown template (leary_html), but currently uses Helvetica font.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A basic example plot",
         subtitle = "A little more detail in substitle?") +
    theme_leary()


}
