% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rprofile.R
\name{update_rprofile}
\alias{update_rprofile}
\title{A function to update my .Rprofile from a version on Github}
\usage{
update_rprofile(
  local.rprofile.path = "~/.Rprofile",
  url = "https://raw.githubusercontent.com/mleary/dotfiles/master/.Rprofile"
)
}
\arguments{
\item{local.rprofile.path}{character value.  Path to local .Rprofile.  Defaults to
commmon "~/.Rprofile".}

\item{url}{character value. URL to RAW github text file with Rprofile template}
}
\description{
A function to update my .Rprofile from a version on Github
}
\details{
This function takes a template Rprofile text file on Github
    (really, this could be URL to a valid file, but Github seems most logical)
    and updates the current working sessions .Rprofile.  Note that it checks
    for the file at the common "~/.Rprofile" location, but can be manually adjusted
    for projects, etc.
}
\examples{
library(learylib)

update_rprofile()  # using default settings

}
