% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leary_presentation.R
\name{leary_presentation}
\alias{leary_presentation}
\title{Custom xarignan presentation template}
\usage{
leary_presentation()
}
\arguments{
\item{None}{Template has all parameters set in the yaml header}
}
\description{
Custom xarignan presentation template
}
\details{
to use the leary_presentation template, simply start a new file and
    select From Tempplate > Leary Presentation
}
