% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_lib_paths.R
\name{see_lib_paths}
\alias{see_lib_paths}
\title{Print the library path(s) from .libPaths()}
\usage{
see_lib_paths()
}
\description{
Print the library path(s) from .libPaths()
}
\details{
This is a helper to let me see the different functions offered by a
   package. NOTE: The package most be installed & loaded (i.e. library(janitor))
}
\examples{
library(learylib)

see_lib_paths() #prints all library path(s)


}
